Move CarouselProject into separate file
